persist: fix idle task ref resolution for negative tids
resolve_task_ref() takes an int tid that is negative for idle tasks
(-(cpu+1)), but the thread-state map is keyed on the u32 wprof_thread.tid
from the register side, which zero-extends. Casting the int straight to
long sign-extends, so negative idle refs never match and every idle task
collapses onto a single tid-0 identity -- one shared per-CPU idle track.
Cast through u32 so the key matches the register side.

+++ src/persist.c
@@ -261,7 +261,7 @@ static int resolve_task_ref(struct persist_state *ps, int tid)
 {
 	struct persist_thread_state *st;
 
-	if (!tid || !hashmap__find(ps->thread_states, (long)tid, &st) || !st->info)
+	if (!tid || !hashmap__find(ps->thread_states, (long)(u32)tid, &st) || !st->info)
 		return 0;
 	if (!st->task_id)
 		st->task_id = persist_task_id(ps, st->info);
